preliminary bug fix: difficulty on created recipe
This fix addresses issue of no difficulty being assigned to a recipe created by a user. It works, and difficulty is now added to a recipe but should circle back and look for a better way.

Redundant code as this defines `calculate_difficulty` function again in views.py. Need to find a way to access `calculate_difficulty` written in models.py